Remove example of quickstart-xgboost-horizontal; move fedxgb_nn_avg from framework to baseline folder
