--- examples/quickstart-xgboost-horizontal/.gitignore
@@ -1,2 +1,0 @@
-dataset
-
